cloud_function_email_handler/requirements.txt
google-cloud-vision

